requirements.txt
arxiv 
google-cloud-texttospeech 
google-cloud-aiplatform
google-cloud-storage
langgraph 
langchain 
langchain-community 
langchain-google-vertexai 
langchain_core 
pymupdf 
wikipedia 
xmltodict 
requests
google-auth
google-cloud
streamlit
ipython
ffmpeg

